+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Renamed `LowerCase` word formatter to `LowerCaseConverter`.
 - Renamed `MixedCase` word formatter to `MixedCaseConverter`.
 - Renamed `UpperCase` word formatter to `UpperCaseConverter`.
+- Renamed `ChangeDate` rule to `ChangedOnDateRule`.
 
 ### Fixed
 - Nothing